refactor: 스크랩 페이지
